+++ 1.21.11/CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v21.11.1-1.21.11] - 2025-12-18
+
+### Changed
+
+- When increasing the item stack size for potions, all potion types are now included, specifically also splash and
+  lingering potions
+- When the larger item stack size is enabled for throwable potions, they have a small use cooldown added
+- Removing the axe attack penalty has been refactored to also include weapons such as swords (in their case the digging
+  penalty is removed)
+
 ## [v21.11.0-1.21.11] - 2025-12-18
 
 ### Changed